Framework: export shared-utils instead of symlinking it
