+++ .gitignore
@@ -20,8 +20,6 @@ env-config.sh
 
 /build
 
-/client/lib/user/shared-utils.js
-
 /public/*.js
 /public/style*.css
 /public/style*.css.map
